requirements.txt
Django==3.2.6
requests==2.26.0
django-crispy-forms
crispy-bootstrap5